NP-2534 Move orcid stuff to orcid component
