+++ src/api/external/orcidApi.ts
@@ -1,11 +1,5 @@
 import Axios from 'axios';
-import { Dispatch } from 'redux';
-
-import { setExternalOrcid } from '../../redux/actions/orcidActions';
-import i18n from '../../translations/i18n';
-import { ORCID_USER_INFO_URL, StatusCode } from '../../utils/constants';
-import { setNotification } from '../../redux/actions/notificationActions';
-import { NotificationVariant } from '../../types/notification.types';
+import { ORCID_USER_INFO_URL } from '../../utils/constants';
 
 interface OrcidInfo {
   id: string;
